requirements.txt
click==8.0.0
colorama==0.4.4
cycler==0.10.0
Flask==2.0.0
greenlet==1.1.0
imbalanced-learn==0.8.0
imblearn==0.0
importlib-metadata==4.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
mysql-connector-python==8.0.25
mysqlclient==2.0.3
numpy==1.20.3
pandas==1.2.4
patsy==0.5.1
Pillow==8.2.0
protobuf==3.17.0
pyarrow==4.0.0
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.11.1
six==1.16.0
sklearn==0.0
SQLAlchemy==1.4.15
statsmodels==0.12.2
threadpoolctl==2.1.0
typing-extensions==3.10.0.0
Werkzeug==2.0.1
xgboost==1.4.2
zipp==3.4.1
